refactor: chain Filter와 nonFilter Chain 구조화
