+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -1,12 +1,11 @@
 package nextstep.auth;
 
-import nextstep.auth.authentication.BasicAuthenticationFilter;
-import nextstep.auth.authentication.BearerTokenAuthenticationFilter;
-import nextstep.auth.authentication.UsernamePasswordAuthenticationFilter;
+import nextstep.auth.authentication.nonchain.UsernamePasswordAuthenticationFilter;
+import nextstep.auth.authentication.chain.AuthenticationChainFilter;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
 import nextstep.auth.context.SecurityContextPersistenceFilter;
 import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenAuthenticationInterceptor;
+import nextstep.auth.authentication.nonchain.TokenAuthenticationInterceptor;
 import nextstep.member.application.LoginMemberService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -18,10 +17,12 @@ import java.util.List;
 public class AuthConfig implements WebMvcConfigurer {
     private LoginMemberService loginMemberService;
     private JwtTokenProvider jwtTokenProvider;
+    private List<AuthenticationChainFilter> chainFilters;
 
-    public AuthConfig(LoginMemberService loginMemberService, JwtTokenProvider jwtTokenProvider) {
+    public AuthConfig(LoginMemberService loginMemberService, JwtTokenProvider jwtTokenProvider, List<AuthenticationChainFilter> chainFilters) {
         this.loginMemberService = loginMemberService;
         this.jwtTokenProvider = jwtTokenProvider;
+        this.chainFilters = chainFilters;
     }
 
     @Override
@@ -29,8 +30,7 @@ public class AuthConfig implements WebMvcConfigurer {
         registry.addInterceptor(new SecurityContextPersistenceFilter());
         registry.addInterceptor(new UsernamePasswordAuthenticationFilter(loginMemberService)).addPathPatterns("/login/form");
         registry.addInterceptor(new TokenAuthenticationInterceptor(loginMemberService, jwtTokenProvider)).addPathPatterns("/login/token");
-        registry.addInterceptor(new BasicAuthenticationFilter(loginMemberService));
-        registry.addInterceptor(new BearerTokenAuthenticationFilter(jwtTokenProvider));
+        chainFilters.forEach(registry::addInterceptor);
     }
 
     @Override
