+++ src/main/java/nextstep/auth/authentication/chain/AuthenticationChainFilter.java
@@ -1,0 +1,6 @@
+package nextstep.auth.authentication.chain;
+
+import org.springframework.web.servlet.HandlerInterceptor;
+
+public interface AuthenticationChainFilter extends HandlerInterceptor {
+}
